requirements.txt
psycopg2==2.9.7
psycopg2-binary==2.9.7
python-dotenv==1.0.0
